feat(table): add uts for the table loading
use shallow instead of shallowWithState if no redux
Rename example for examples
Extract the PaginationSelectors
