+++ packages/react-vapor/src/components/loading/components/tests/ActionBarLoading.spec.tsx
@@ -1,4 +1,4 @@
-import {shallowWithState} from 'enzyme-redux';
+import {shallow} from 'enzyme';
 import * as React from 'react';
 import {ActionBarLoading} from '../ActionBarLoading';
 
@@ -6,7 +6,7 @@ describe('ActionBarLoading tests', () => {
     describe('<ActionBarLoading />', () => {
         it('should mount and unmount without errors', () => {
             expect(() => {
-                const wrapper = shallowWithState(<ActionBarLoading />, {});
+                const wrapper = shallow(<ActionBarLoading />, {});
                 wrapper.unmount();
             });
         });
