+++ packages/react-vapor/src/components/loading/components/tests/BasicHeaderLoading.spec.tsx
@@ -1,4 +1,4 @@
-import {shallowWithState} from 'enzyme-redux';
+import {shallow} from 'enzyme';
 import * as React from 'react';
 import {BasicHeaderLoading} from '../BasicHeaderLoading';
 
@@ -6,7 +6,7 @@ describe('BasicHeaderLoading tests', () => {
     describe('<BasicHeaderLoading />', () => {
         it('should mount and unmount without errors', () => {
             expect(() => {
-                const wrapper = shallowWithState(<BasicHeaderLoading />, {});
+                const wrapper = shallow(<BasicHeaderLoading />, {});
                 wrapper.unmount();
             });
         });
