+++ packages/react-vapor/src/components/loading/components/tests/ContentLoadingPlaceholder.spec.tsx
@@ -1,4 +1,4 @@
-import {shallowWithState} from 'enzyme-redux';
+import {shallow} from 'enzyme';
 import * as React from 'react';
 import {ContentLoadingPlaceholder} from '../ContentLoadingPlaceholder';
 
@@ -6,7 +6,7 @@ describe('ContentLoadingPlaceholder tests', () => {
     describe('<ContentLoadingPlaceholder />', () => {
         it('should mount and unmount without errors', () => {
             expect(() => {
-                const wrapper = shallowWithState(
+                const wrapper = shallow(
                     <ContentLoadingPlaceholder>
                         <div>Test</div>
                     </ContentLoadingPlaceholder>,
