+++ packages/react-vapor/src/components/loading/components/tests/PaginationLoading.spec.tsx
@@ -1,4 +1,4 @@
-import {shallowWithState} from 'enzyme-redux';
+import {shallow} from 'enzyme';
 import * as React from 'react';
 import {PaginationLoading} from '../PaginationLoading';
 
@@ -6,7 +6,7 @@ describe('PaginationLoading tests', () => {
     describe('<PaginationLoading />', () => {
         it('should mount and unmount without errors', () => {
             expect(() => {
-                const wrapper = shallowWithState(<PaginationLoading />, {});
+                const wrapper = shallow(<PaginationLoading />, {});
                 wrapper.unmount();
             });
         });
