+++ packages/react-vapor/src/components/loading/components/tests/TableLoading.spec.tsx
@@ -1,4 +1,4 @@
-import {shallowWithState} from 'enzyme-redux';
+import {shallow} from 'enzyme';
 import * as React from 'react';
 import {TableLoading} from '../TableLoading';
 
@@ -6,7 +6,7 @@ describe('TableLoading tests', () => {
     describe('<TableLoading.Table />', () => {
         it('should mount and unmount without errors', () => {
             expect(() => {
-                const wrapper = shallowWithState(<TableLoading.Table />, {});
+                const wrapper = shallow(<TableLoading.Table />, {});
                 wrapper.unmount();
             });
         });
@@ -15,18 +15,18 @@ describe('TableLoading tests', () => {
     describe('<TableLoading.Body />', () => {
         it('should mount and unmount without errors', () => {
             expect(() => {
-                const wrapper = shallowWithState(<TableLoading.Body />, {});
+                const wrapper = shallow(<TableLoading.Body />, {});
                 wrapper.unmount();
             });
         });
 
         it('should render <tr/> equal of the the number of columns sent as parameter', () => {
-            const wrapper = shallowWithState(<TableLoading.Body numberOfRow={10} />, {});
+            const wrapper = shallow(<TableLoading.Body numberOfRow={10} />, {});
             expect(wrapper.find('tr').length).toBe(10);
         });
 
         it('should render <Row/> equal of the the number of columns sent as parameter', () => {
-            const wrapper = shallowWithState(<TableLoading.Body numberOfColumns={8} numberOfRow={1} />, {});
+            const wrapper = shallow(<TableLoading.Body numberOfColumns={8} numberOfRow={1} />, {});
             expect(wrapper.find(TableLoading.Row).length).toBe(8);
         });
     });
@@ -34,18 +34,18 @@ describe('TableLoading tests', () => {
     describe('<TableLoading.Row />', () => {
         it('should mount and unmount without errors', () => {
             expect(() => {
-                const wrapper = shallowWithState(<TableLoading.Row num={0} />, {});
+                const wrapper = shallow(<TableLoading.Row num={0} />, {});
                 wrapper.unmount();
             });
         });
 
         it('should add the class mod-haft if the number is odd', () => {
-            const wrapper = shallowWithState(<TableLoading.Row num={1} />, {});
+            const wrapper = shallow(<TableLoading.Row num={1} />, {});
             expect(wrapper.find('div').hasClass('mod-half')).toBe(true);
         });
 
         it('should not add the class mod-haft if the number is even', () => {
-            const wrapper = shallowWithState(<TableLoading.Row num={2} />, {});
+            const wrapper = shallow(<TableLoading.Row num={2} />, {});
             expect(wrapper.find('div').hasClass('mod-half')).toBe(false);
         });
     });
