+++ packages/react-vapor/src/components/pagination/index.ts
@@ -2,4 +2,5 @@ export * from './PaginationPagesNumber';
 export * from './PaginationPerPage';
 export * from './PaginationSelect';
 export * from './PaginationUtils';
+export * from './PaginationSelectors';
 export * from './TablePagination';
