+++ packages/react-vapor/src/components/pagination/tests/PaginationPerPage.spec.tsx
@@ -1,4 +1,4 @@
-import {shallowWithState} from 'enzyme-redux';
+import {shallow} from 'enzyme';
 import * as React from 'react';
 import {FlatSelectWithPrepend} from '../../flatSelect/FlatSelectWithPrepend';
 import {PaginationPerPage} from '../PaginationPerPage';
@@ -7,13 +7,13 @@ describe('PaginationPerPage', () => {
     describe('<PaginationPerPage /> tests', () => {
         it('should mount and unmount without errors', () => {
             expect(() => {
-                const wrapper = shallowWithState(<PaginationPerPage id={'id'} />, {});
+                const wrapper = shallow(<PaginationPerPage id={'id'} />, {});
                 wrapper.unmount();
             });
         });
 
         it('should add an option for each number in the perPage array', () => {
-            const wrapper = shallowWithState(<PaginationPerPage id={'id'} perPage={[1, 2, 3]} />, {});
+            const wrapper = shallow(<PaginationPerPage id={'id'} perPage={[1, 2, 3]} />, {});
             expect(wrapper.find(FlatSelectWithPrepend).props().options.length).toBe(3);
         });
     });
