+++ packages/react-vapor/src/components/pagination/tests/TablePagination.spec.tsx
@@ -1,4 +1,4 @@
-import {shallowWithState} from 'enzyme-redux';
+import {shallow} from 'enzyme';
 import * as React from 'react';
 import {PaginationPerPage} from '../PaginationPerPage';
 import {TablePagination} from '../TablePagination';
@@ -7,7 +7,7 @@ describe('TablePagination', () => {
     describe('<TablePagination /> tests', () => {
         it('should mount and unmount without errors', () => {
             expect(() => {
-                const wrapper = shallowWithState(
+                const wrapper = shallow(
                     <TablePagination
                         id={'id'}
                         totalPages={1}
@@ -24,7 +24,7 @@ describe('TablePagination', () => {
 
         describe('<PaginationPerPage /> child', () => {
             it('should set <PaginationPerPage/> hidden if the perPageNumbers is empty', () => {
-                const wrapper = shallowWithState(
+                const wrapper = shallow(
                     <TablePagination
                         id={'id'}
                         totalPages={1}
@@ -39,7 +39,7 @@ describe('TablePagination', () => {
             });
 
             it('should set <PaginationPerPage/> hidden if the first value of perPageNumber is greater than the totalEntries', () => {
-                const wrapper = shallowWithState(
+                const wrapper = shallow(
                     <TablePagination
                         id={'id'}
                         totalPages={1}
