+++ packages/react-vapor/src/components/table-hoc/examples/TableHOCServerExamples.tsx
@@ -5,39 +5,33 @@ import {connect} from 'react-redux';
 import {RouteComponentProps, withRouter} from 'react-router';
 import * as _ from 'underscore';
 
-import {ExampleComponent} from '../../../../docs/src/components/ComponentsInterface';
 import {withServerSideProcessing} from '../../../hoc/withServerSideProcessing/withServerSideProcessing';
-import {DateUtils} from '../../../utils/DateUtils';
-import {IDispatch, IReduxAction, IThunkAction} from '../../../utils/ReduxUtils';
+import {IDispatch, IThunkAction} from '../../../utils/ReduxUtils';
 import {IReactVaporTestState} from '../../../utils/tests/TestUtils';
-import {SELECTION_BOXES_LONG} from '../../datePicker/examples/DatePickerExamplesCommon';
+import {filterThrough} from '../../filterBox';
 import {LastUpdated} from '../../lastUpdated/LastUpdated';
-import {turnOffLoading} from '../../loading/LoadingActions';
 import {Section} from '../../section/Section';
 import {TableWithPaginationActions} from '../actions/TableWithPaginationActions';
 import {TableHeaderWithSort} from '../TableHeaderWithSort';
 import {TableHOC} from '../TableHOC';
-import {ITableHOCCompositeState, TableHOCUtils} from '../TableHOCUtils';
-import {TableRowConnected} from '../TableRowConnected';
-import {TableRowNumberColumn} from '../TableRowNumberColumn';
+import {TableRowHeader} from '../TableRowHeader';
 import {TableRowNumberHeader} from '../TableRowNumberHeader';
 import {tableWithActions} from '../TableWithActions';
 import {tableWithBlankSlate} from '../TableWithBlankSlate';
 import {tableWithDatePicker} from '../TableWithDatePicker';
 import {tableWithFilter} from '../TableWithFilter';
-import {tableWithPagination} from '../TableWithPagination';
+import {tableWithNewPagination} from '../TableWithNewPagination';
 import {tableWithPredicate} from '../TableWithPredicate';
 import {tableWithSort} from '../TableWithSort';
 import {tableWithUrlState} from '../TableWithUrlState';
+import {TableHOCExampleUtils, TableHOCServerExampleContext} from '../utils/TableHOCExampleUtils';
+import {ITableHOCCompositeState, TableHOCUtils} from '../utils/TableHOCUtils';
 
-type TableHOCServerProps = RouteComponentProps &
-    ReturnType<typeof mapStateToProps> &
-    ReturnType<typeof mapDispatchToProps>;
+type TableHOCServerProps = RouteComponentProps & ReturnType<typeof mapDispatchToProps>;
 
 export interface IExampleRowData {
     city: string;
     email: string;
-
     username: string;
     dateOfBirth: Date;
     id: string;
@@ -48,177 +42,138 @@ export interface IExampleServerTableState {
     isLoading: boolean;
 }
 
-interface ISetExampleDataPayload {
-    data: any[];
-}
-
-interface ISetExampleIsLoadingPayload {
+interface TableHOCServerExamplesState {
+    data: {users: [any]; count: number};
     isLoading: boolean;
 }
 
-type IExamplePayload = ISetExampleDataPayload | ISetExampleIsLoadingPayload;
-
-export const TableHOCServerExamples: ExampleComponent = () => <TableHOCServer />;
+export const TableHOCServerExamples = () => <TableHOCServer />;
 
 TableHOCServerExamples.title = 'TableHOC server';
 
 // start-print
 export const TableHOCServerExampleId = 'complex-example';
 
-const tableActions = (username: string) => [
-    {
-        primary: true,
-        icon: 'edit',
-        name: 'edit',
-        enabled: true,
-        trigger: () => alert(username),
-        callOnDoubleClick: true,
-    },
-];
-
-const generateRows = (allData: IExampleRowData[]) =>
-    allData.map((data: IExampleRowData, i: number) => (
-        <TableRowConnected
-            id={data.username}
-            tableId={TableHOCServerExampleId}
-            key={data.username}
-            actions={tableActions(data.username)}
-            isMultiselect
-            disabled={i % 3 === 0}
-            collapsible={{content: <div className="py2">👋</div>}}
-        >
-            <TableRowNumberColumn number={i + 1} />
-            <td key="city">{data.city}</td>
-            <td key="email">{data.email.toLowerCase()}</td>
-            <td key="username">{data.username.toLowerCase()}</td>
-            <td key="date-of-birth">{data.dateOfBirth.toLocaleDateString()}</td>
-        </TableRowConnected>
-    ));
-
 const renderHeader = () => (
-    <thead>
-        <tr>
-            <TableRowNumberHeader />
-            <TableHeaderWithSort id="address.city" tableId={TableHOCServerExampleId}>
-                City
-            </TableHeaderWithSort>
-            <TableHeaderWithSort id="email" tableId={TableHOCServerExampleId}>
-                Email
-            </TableHeaderWithSort>
-            <TableHeaderWithSort id="username" tableId={TableHOCServerExampleId} isDefault>
-                Username
-            </TableHeaderWithSort>
-            <th key="date-of-birth">Date of Birth</th>
-            <th>{/* Empty th for the collapsible */}</th>
-        </tr>
-    </thead>
+    <TableHOCServerExampleContext.Consumer>
+        {({isLoading}) => (
+            <thead>
+                <tr>
+                    <TableRowNumberHeader isLoading={isLoading} />
+                    <TableHeaderWithSort id="address.city" tableId={TableHOCServerExampleId} isLoading={isLoading}>
+                        City
+                    </TableHeaderWithSort>
+                    <TableHeaderWithSort id="email" tableId={TableHOCServerExampleId} isLoading={isLoading}>
+                        Email
+                    </TableHeaderWithSort>
+                    <TableHeaderWithSort
+                        id="username"
+                        tableId={TableHOCServerExampleId}
+                        isLoading={isLoading}
+                        isDefault
+                    >
+                        Username
+                    </TableHeaderWithSort>
+                    <TableRowHeader isLoading={isLoading}>Date of Birth</TableRowHeader>
+                    <TableRowHeader isLoading={isLoading} />
+                </tr>
+            </thead>
+        )}
+    </TableHOCServerExampleContext.Consumer>
 );
 
-const tablePredicates = [
-    {
-        id: 'address.city',
-        prepend: <span className="mr1 text-medium-grey">City:</span>,
-        values: [
-            {displayValue: 'All', value: '', selected: true},
-            {displayValue: 'Lebsackbury', value: 'Lebsackbury'},
-        ],
-    },
-    {
-        id: 'username',
-        prepend: <span className="mr1 text-medium-grey">Username:</span>,
-        values: [
-            {displayValue: 'All', value: '', selected: true},
-            {displayValue: 'bret', value: 'Bret'},
-        ],
-    },
-];
-
-const tableDatePickerConfig = {
-    datesSelectionBoxes: SELECTION_BOXES_LONG,
-    years: [...DateUtils.getPreviousYears(25), DateUtils.currentYear.toString()],
-    initialDateRange: [
-        moment()
-            .subtract(25, 'years')
-            .toDate(),
-        moment().toDate(),
-    ],
-};
-
-const mapStateToProps = (state: IReactVaporTestState) => ({
-    isLoading: state.tableHOCExample.isLoading,
-    serverData: state.tableHOCExample.data,
-});
 const mapDispatchToProps = (dispatch: IDispatch) => ({
-    fetch: _.debounce(() => dispatch(TableHOCServerActions.fetchData()), 400),
+    fetch: () => dispatch(TableHOCServerActions.fetchData()),
+    resetFilter: () => dispatch(filterThrough(TableHOCServerExampleId, '')),
 });
 
-const TableExampleDisconnected: React.FunctionComponent<TableHOCServerProps> = (props) => {
-    const onUpdate = () => {
-        props.fetch();
+class TableExampleDisconnected extends React.PureComponent<TableHOCServerProps, TableHOCServerExamplesState> {
+    state: TableHOCServerExamplesState = {
+        data: null,
+        isLoading: true,
     };
 
-    const updateUrl = (query: string) => {
-        props.history.push({search: query});
+    private ServerTableComposed = _.compose(
+        withServerSideProcessing,
+        tableWithUrlState,
+        tableWithBlankSlate({
+            title: 'No data fetched from the server',
+            description: 'Try reviewing the specified filters above or clearing all filters.',
+            buttons: [
+                {
+                    name: 'Clear filter',
+                    enabled: true,
+                    onClick: () => this.props.resetFilter(),
+                },
+            ],
+        }),
+        tableWithPredicate(TableHOCExampleUtils.tablePredicates[0]),
+        tableWithPredicate(TableHOCExampleUtils.tablePredicates[1]),
+        tableWithFilter({
+            placeholder: 'Filter all',
+            blankSlate: {
+                title: 'No results found',
+            },
+        }),
+        tableWithSort(),
+        tableWithDatePicker({...(TableHOCExampleUtils.tableDatePickerConfig as any)}),
+        tableWithNewPagination({perPageNumbers: [3, 5, 10]}),
+        tableWithActions()
+    )(TableHOC);
+
+    private fetch = _.debounce(() => {
+        this.setState({...this.state, isLoading: true});
+        window.setTimeout(
+            () =>
+                this.props.fetch().done((data: any) => {
+                    this.setState({data, isLoading: false});
+                }),
+            500
+        );
+    }, 40);
+
+    private onUpdate = () => {
+        this.fetch();
     };
 
-    return (
-        <Section title="Server table with numbered rows">
-            <span className="block my2 text-grey-7">
-                Please note that the backend service doesn't support dates but we still make a request for every change
-                in the date range.
-            </span>
-            <ServerTableComposed
-                id={TableHOCServerExampleId}
-                className="table table-numbered mod-collapsible-rows"
-                data={props.serverData}
-                renderBody={generateRows}
-                tableHeader={renderHeader()}
-                onUpdate={onUpdate}
-                onUpdateUrl={updateUrl}
-                isLoading={props.isLoading}
-            >
-                <LastUpdated time={new Date()} />
-            </ServerTableComposed>
-        </Section>
-    );
-};
+    private updateUrl = (query: string) => {
+        this.props.history.push({search: query});
+    };
 
-const ServerTableComposed = _.compose(
-    withServerSideProcessing,
-    tableWithUrlState,
-    tableWithBlankSlate({title: 'No data fetched from the server'}),
-    tableWithPredicate({...tablePredicates[0]}),
-    tableWithPredicate({...tablePredicates[1]}),
-    tableWithFilter(),
-    tableWithDatePicker({...(tableDatePickerConfig as any)}),
-    tableWithSort(),
-    tableWithPagination({perPageNumbers: [3, 5, 10]}),
-    tableWithActions()
-)(TableHOC);
+    componentDidMount() {
+        this.fetch();
+    }
 
-const TableHOCServer = connect(
-    mapStateToProps,
+    render() {
+        return (
+            <Section title="Server table with numbered rows">
+                <span className="block my2 text-grey-7">
+                    Please note that the backend service doesn't support dates but we still make a request for every
+                    change in the date range.
+                </span>
+                <TableHOCServerExampleContext.Provider
+                    value={{isLoading: this.state.isLoading, id: TableHOCServerExampleId}}
+                >
+                    <this.ServerTableComposed
+                        id={TableHOCServerExampleId}
+                        className="table table-numbered mod-collapsible-rows"
+                        data={this.state.data?.users ?? []}
+                        renderBody={TableHOCExampleUtils.generateRows}
+                        tableHeader={renderHeader()}
+                        onUpdate={this.onUpdate}
+                        onUpdateUrl={this.updateUrl}
+                        isLoading={this.state.isLoading}
+                        loading={{numberOfColumns: 6}}
+                    >
+                        <LastUpdated time={new Date()} />
+                    </this.ServerTableComposed>
+                </TableHOCServerExampleContext.Provider>
+            </Section>
+        );
+    }
+}
 
-    mapDispatchToProps
-)(withRouter(TableExampleDisconnected));
-
-/* ACTIONS */
-
-export const TableHOCServerActionsType = {
-    setData: 'TABLE_HOC_SET_DATA',
-    setIsLoading: 'TABLE_HOC_SET_IS_LOADING',
-    fetch: 'TABLE_HOC_FETCH_DATA',
-};
-
-const setData = (data: any[]): IReduxAction<ISetExampleDataPayload> => ({
-    type: TableHOCServerActionsType.setData,
-    payload: {data},
-});
-
-const setIsLoading = (isLoading: boolean): IReduxAction<ISetExampleIsLoadingPayload> => ({
-    type: TableHOCServerActionsType.setIsLoading,
-    payload: {isLoading},
-});
+const TableHOCServer = connect(undefined, mapDispatchToProps)(withRouter(TableExampleDisconnected));
 
 const fetchData = (): IThunkAction => (dispatch: IDispatch, getState: () => IReactVaporTestState) => {
     const compositeState: ITableHOCCompositeState = TableHOCUtils.getCompositeState(
@@ -238,10 +193,9 @@ const fetchData = (): IThunkAction => (dispatch: IDispatch, getState: () => IRea
     _.each(compositeState.predicates, (predicate: {id: string; value: string}) => {
         params[predicate.id] = predicate.value;
     });
-    dispatch(setIsLoading(true));
-    $.get('https://jsonplaceholder.typicode.com/users', params).done((response: any[], status, request) => {
+    return $.get('https://jsonplaceholder.typicode.com/users', params).then((response: any[], status, request) => {
         const count = request.getResponseHeader('x-total-count');
-        const users = response.map((user: any) => ({
+        const users = _.map(response, (user: any) => ({
             city: user.address.city,
             username: user.username,
             email: user.email,
@@ -249,38 +203,14 @@ const fetchData = (): IThunkAction => (dispatch: IDispatch, getState: () => IRea
                 .subtract(user.address.city.length, 'years')
                 .toDate(), // fake a year of birth
         }));
-        dispatch(setData(users));
-        dispatch(turnOffLoading([TableHOCServerExampleId]));
-        dispatch(TableWithPaginationActions.setCount(TableHOCServerExampleId, count));
+        dispatch(TableWithPaginationActions.setCount(TableHOCServerExampleId, count as any));
+        return {
+            count,
+            users,
+        };
     });
 };
 
-export const TableHOCServerActions = {
-    setData,
-    setIsLoading,
+const TableHOCServerActions = {
     fetchData,
-};
-
-/* REDUCER */
-
-export const TableHOCServerExampleReducer = (
-    state: IExampleServerTableState = {data: [], isLoading: true},
-    action: IReduxAction<IExamplePayload>
-) => {
-    if (action.type === TableHOCServerActionsType.setData) {
-        const payload = action.payload as ISetExampleDataPayload;
-        return {
-            ...state,
-            data: [...payload.data],
-            isLoading: false,
-        };
-    }
-    if (action.type === TableHOCServerActionsType.setIsLoading) {
-        const payload = action.payload as ISetExampleIsLoadingPayload;
-        return {
-            ...state,
-            isLoading: payload.isLoading,
-        };
-    }
-    return state;
 };
